+++ template/c-bindings/template/lib/bindings.ml
@@ -1,0 +1,1 @@
+include Demo_stubs.Def (Demo_generated_stubs)