+++ template/c-bindings/template/lib/stubs/{{ project_snake }}_stubs.ml
@@ -1,0 +1,7 @@
+open Ctypes
+
+module Def (F : Cstubs.FOREIGN) = struct
+  open F
+
+  let printf = F.foreign "printf" (string @-> returning void)
+end