+++ template/c-bindings/template/lib/utils.ml
@@ -1,0 +1,3 @@
+let printf s = Bindings.printf s
+
+let greet name = "Hello " ^ name ^ "!"