+++ test_bin/success/list-official-templates.t
@@ -3,6 +3,9 @@
     bin
       Native project containing a binary
   
+    c-bindings
+      Bindings to a C library
+  
     cli
       Command Line Interface releasable on Opam